requirements.txt
Django
psycopg2
django-seed
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers


